HV: add support to boot SOS with OVMF
Add support to boot SOS with OVMF. Use bootloaders that support multiboot/multiboot2 (e.g.grub) to boot acrn.32.out with a modified version of OVMF (https://github.com/qianwang-intel/acrn-edk2 branch:ovmf-acrn) as the module/module2.

To boot SOS via OVMF, OVMF image will be loaded as-is into memory and the address will be passed to hypervisor within the multiboot header.
HV will remap the memory segment containing OVMF image to the tail of GPA 4G in EPT, and reset the vcpu to real mode and set ip=0xFFF0.

This method is tested on whl-ipc-i5 and kbl-nuc-i7 platform.